backend/feat/BKN-2984/Added-DriverSpeed-inPool

+++ Backend/app/provider-platform/dynamic-offer-driver-app/Main/src/Domain/Types/TransporterConfig.hs
@@ -46,7 +46,7 @@ data TransporterConfigD u = TransporterConfig
     driverSpeedWeightage :: Int,
     minLocationUpdates :: Int,
     locationUpdateSampleTime :: Minutes,
-    defaultDriverSpeed :: Int
+    defaultDriverSpeed :: Double
   }
   deriving (Generic, Show)
